+++ app/src/stock.py
@@ -7,13 +7,19 @@ from product import Product
 class Stock:
     products: Dict[Product, int] = field(default_factory=dict)
 
-    def add_product(self, product, quantity):
+    def remove_product(self, product, quantity=1):
         if product in self.products:
-            self.products[product] += quantity
-        else:
-            self.products[product] = quantity
+            if self.products[product] >= quantity:
+                self.products[product] -= quantity
+                if self.products[product] == 0:
+                    del self.products[product]
+            else:
+                raise ValueError("Insufficient stock")
 
-    def remove_product(self, product, quantity):
+    def remove_product(self, product, quantity=1):
+        if not isinstance(quantity, int) or quantity <= 0:
+            raise ValueError("Quantity must be a positive integer")
+
         if product in self.products:
             if self.products[product] >= quantity:
                 self.products[product] -= quantity
@@ -27,28 +33,3 @@ class Stock:
 
     def check_stock_availability(self, product, quantity):
         return self.get_stock_quantity(product) >= quantity
-
-
-# Example usage:
-
-# Create some products
-product1 = Product(
-    1,
-    "Product 1",
-    "Description of Product 1",
-    19.99,
-    "product1.jpg",
-    stock_quantity=50,
-    category="Electronics",
-)
-product2 = Product(
-    2,
-    "Product 2",
-    "Description of Product 2",
-    29.99,
-    "product2.jpg",
-    stock_quantity=100,
-    category="Clothing",
-)
-
-#
